Removing O_APPEND from output file
This simple mistake was making simulation in SAMPLER into a nightmare as
it's catch-up routine was constantly failing due to format-change of
previous entries. We also need to be able simulate the very real case of
files actually stopping to exist or being renamed wile reading the. If
O_APPEND is going to be used it should be made optional via a flag.