Crude user-interactivity deployed
Crude framework. Can currently only start and top playback of file ATM. This
commit is a proof of concept. Most if not all is contained in new file(s)
outside of current HEMUL scope.

Future features could be:

* Entering command mode (-o and -i flag deployed)
* Changing interactive flow via -O and -I flags (stdin/stdout files). By
  rerouting these to prepared named pipes, control over tcp can be deployed
  with ease.
* With or without stopping replay first
* Shortcuts (keys) for operation without feedback
   1) Toggling or
   2) setting arguments that affect flow
   I.e. period-time, buffering
* Functions of interest.
   1) Flush on demand
   2) Playing backwards
   3) Changing buffer-size
   4) Composing specific buffer out of lines

Need curses for this? Hopefully not, or transparency over tcp will likely
not work.

References: man struct termios