Bug: When reading decimal timestamps.
scanf failed when type-cast int values were used. This is somewhat strange as
values should really only had been truncated. The error affected only tv_sec
part and for time-stamps like 177.165008, tv_sec would be scanned to a
incorrect high value. tv_usec was always read correctly.

This can only be explained by the fact that these two have different types.
The solution of not typecasting and increasing size for each to print-long
implies that they are nevertheless the same size as gcc doesn't complain.

It is unfortunate that printf/scanf with friends cant be specified more
precisely as both the size of formatters and the size of struct timeval
members can change. To do this better, one should have a dedicated function
whith pre-processor conditionals. It's good for now however.

This error was found by accident which failing to specify a good regexp for
kernel log time-stamps. The following one turns out as recommended as a
minimum:

-R'^.*\[([[:space:]]*[[:digit:]]+\.[[:digit:]]+)\].*'

Notice the anchor, '[:space:]' and the two '+' & '\[' '\]'. These are
important for a general pattern, but that is specific enough. The full
pattern for a kernel-log time-stamp however is:

^<[0-9]>\[[[:space:]]*([[:digit:]]+\.[[:digit:]]+)\].*